// src/lib/api/generated.ts
export interface paths {
// ...
    "/audits": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * All
         * @description Role: CTO
         */
        get: {
// ...
            responses: {
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
// ...
                        "application/json": {
                            data: {
                                current_page: number;
                                data: {
                                    id: number;
                                    user_id: number;
                                    user_name: string;
                                    user_role: string;
                                    event: string;
                                    auditable_type: string;
                                    auditable_id: number;
                                    created_at: string;
                                    updated_at: string;
                                    old_values: string[];
                                    new_values: {
                                        user_id: number;
                                        status: string;
                                        notes: string | null;
                                        id: number;
                                        first_name?: string;
                                        last_name?: string;
                                        username?: string;
                                        password?: string;
                                        department_id?: number;
                                    };
                                }[];
                                first_page_url: string;
                                from: number;
                                last_page: number;
                                last_page_url: string;
                                links: {
                                    url: string | null;
                                    label: string;
                                    page: number | null;
                                    active: boolean;
                                }[];
                                next_page_url: null;
                                path: string;
                                per_page: number;
                                prev_page_url: null;
                                to: number;
                                total: number;
                            };
                            message: string;
                        };
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}